// src/types/auth/googleSignup/body.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface GoogleSignupBody {
  access_token: string;
  user_data: {
    age?: number;
    gender?: string;
    [k: string]: unknown;
  };
  [k: string]: unknown;
}